+++ src/components/AddonsList.vue
@@ -137,16 +137,6 @@
     >
       <div class="avatar">
         <div class="w-12 rounded">
-          <img src="/public/logos/cometfr.png" alt="CometFR" />
-        </div>
-      </div>
-      <p class="text-sm mt-2 text-center">CometFR</p>
-    </div>
-    <div
-      class="flex flex-col items-center p-4 hover:bg-base-200 rounded-lg transition-colors"
-    >
-      <div class="avatar">
-        <div class="w-12 rounded">
           <img src="/public/logos/brazucatorrents.png" alt="Brazuca Torrents" />
         </div>
       </div>